feat: Add userinfo validation (#18)
